Decompose BoundaryConditionsPanel into focused hooks/components (issue #346)
Split the 841-line monolith into the same shape as the LeftPanel
decomposition (#165/#327): shared state in a hook, small focused
components, and a thin composing panel.

- hooks/usePickedFaces.ts — shared face-picking session state
- panel/bcFormUtils.ts — DOF/force/moment labels + parse/format helpers
- panel/BcLoadFormControls.tsx — PickedFaceList, DofCheckboxes,
  LoadKindSelect, LoadVectorInputs, PressureInput
- panel/GroupValueForms.tsx — the two inline ✎ editors (BcValueForm,
  LoadValueForm)
- panel/GroupCard.tsx — shared group-list card, deduplicating the
  previously copy-pasted BC/load group list rendering
- panel/BcSection.tsx / panel/LoadSection.tsx — the two sections
- BoundaryConditionsPanel.tsx — thin composition (39 lines)

Behavior-preserving: all DOM structure, titles, and test ids unchanged;
edit-constraint-values.spec.ts passes.
